requirements.txt
fastapi
uvicorn
pandas
python-dotenv
ccxt
alpaca-py
dukascopy-python
